qualifiers/challenges/ropably/solution/requirements.txt
angr
capstone
keystone-engine
lief
pwntools
